Updated file structure and ci/cd pipeline
