+++ .github/workflows/pytest.yml
@@ -29,6 +29,18 @@ jobs:
         pip install -r requirements.txt
         pip install -e .
 
+    - name: Create binary directory
+      run: |
+        mkdir build_gs
+        cd build_gs
+        cmake .. -DCMAKE_BUILD_TYPE=GS
+
+    - name: Build
+      run: |
+        cd build_gs
+        cmake .. -DCMAKE_BUILD_TYPE=GS
+        cmake --build .
+
     - name: Run pytest
       run: |
         python -m pytest
